refactor: changing the name of src collection module to avoid git and dockerinore issues
