+++ evaluation/run_evaluation.py
@@ -1,6 +1,6 @@
 import os
 
-from src.collection.evaluate_collection import (
+from src.collection_utils.evaluate_collection import (
     assess_retrieval_accuracy,
     get_data_for_evaluation,
     get_all_regex_ids,
